requirements.txt
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==3.0.4
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
coverage==7.6.12
cryptography==44.0.2
daphne==3.0.2
Django==5.1.6
django-admin==2.0.2
django-channels==0.7.0
django-excel-response2==3.0.6
django-six==1.0.5
djangorestframework==3.15.2
excel-base==1.0.4
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
isoweek==1.3.3
msgpack==1.1.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-asyncio==0.25.3
pytest-cov==6.0.0
pytest-django==4.10.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
screen==1.0.1
service-identity==24.2.0
setuptools==75.8.2
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0
xlwt==1.3.0
zope.interface==7.2
